Fix #382 #383 arrumando duplicações
